Enables saving playlists to the database
Adds functionality for users to save their created playlists to the database.

This includes creating a new `DatabasePlaylist` model and establishing relationships with the `Song` and `User` models.

The `PlaylistService` is updated to include a `SavePlaylistToDatabase` method,
which persists the current session playlist to the database, associating it with the logged-in user.

A new action `SavePlaylist` is added to the `PlaylistController` to trigger this save operation.